+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
+
 ## 2.17.0
 * Update @braintree/wrap-promise to v2.0.0
   - Errors thrown inside developer provided callback functions will log in the console instead of killing the node process [wrap-promise#4](https://github.com/braintree/wrap-promise/issues/4)